Separate runner in pre_runner, runner and post_runner
